feat: rename all slidePane to drawer and rename switchWindow folder to useWindowSwitchListener
