+++ src/drawer/style.scss
@@ -1,4 +1,4 @@
-$prefix: "dtc-slide-pane";
+$prefix: "dtc-drawer";
 
 .#{$prefix} {
     top: 0;
