+++ src/index.ts
@@ -5,7 +5,7 @@ export { default as ContentLayout } from './contentLayout';
 export { default as ContextMenu } from './contextMenu';
 export { default as useCookieListener } from './cookies';
 export { default as Copy } from './copy';
-export { default as SlidePane } from './drawer';
+export { default as Drawer } from './drawer';
 export { default as Dropdown } from './dropdown';
 export { default as EllipsisText } from './ellipsisText';
 export { default as Empty } from './empty';
@@ -26,10 +26,10 @@ export { default as ProgressLine } from './progressLine';
 export { default as Resize } from './resize';
 export { default as SpreadSheet } from './spreadSheet';
 export { default as StatusTag } from './statusTag';
-export { default as useWindowSwitchListener } from './switchWindow';
 export { default as Table } from './table';
 export { default as TinyTag } from './tinyTag';
 export { default as useDebounce } from './useDebounce';
 export { default as useIntersectionObserver } from './useIntersectionObserver';
 export { default as useList } from './useList';
 export { default as useModal } from './useModal';
+export { default as useWindowSwitchListener } from './useWindowSwitchListener';
